requirements.txt
qasgiref==3.6.0
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
future==0.18.3
pymysql==1.1.0
pycryptodome==3.18.0
PyJWT==2.7.0
pytz==2023.3
serializers==0.2.4
sqlparse==0.4.4
tzdata==2023.3
